Fix TypeScript type of optional custom cache
p-memoize takes a function that returns promised values, and may optionally
take a custom cache to store these promised values. However, the type
of the custom cache was incorrectly "ReturnType" instead of "PromiseLike<ReturnType>",
meaning if you create a correctly typed cache that stores promised values, it will
not be possible to pass it to p-memoize without any-casting it. Instead, p-memoize
expects a cache of resolved values, which is not what it actually is.